+++ clojure/src/test/java/cucumber/runtime/clojure/ClojureSnippetTest.java
@@ -22,7 +22,7 @@ public class ClojureSnippetTest {
         String expected = "" +
                 "(Given #\"^I have (\\d+) cukes in my \\\"([^\\\"]*)\\\" belly$\" [arg1 arg2]\n" +
                 "  (comment  Express the Regexp above with the code you wish you had  )\n" +
-                "  (throw (cucumber.runtime.PendingException.)))\n";
+                "  (throw (cucumber.api.PendingException.)))\n";
         assertEquals(expected, snippet);
     }
 
@@ -34,7 +34,7 @@ public class ClojureSnippetTest {
         String expected = "" +
                 "(Given #\"^I have:$\" [arg1]\n" +
                 "  (comment  Express the Regexp above with the code you wish you had  )\n" +
-                "  (throw (cucumber.runtime.PendingException.)))\n";
+                "  (throw (cucumber.api.PendingException.)))\n";
         assertEquals(expected, snippet);
     }
 }